community/discord-bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
feedparser>=6.0.0
psutil>=5.9.0